Change message listener to be an object
messageParser dot js used to export a single handle message function. now it
exports a MessageParser object.

listeners can be added to the message parser with a method.

listeners have been moved to lib/messageListeners.

the hope is that logging will get a little easier and adding listeners will as
well. i'm not sure on these names though...