Improve how player stats are reported to slack
Previously, messages wouldn't be posted until we had received data from the nba
api. This led users to wonder if it had worked at all and stats could pop up
deeper in a conversation thread.

Now the app will post that it is looking for stats for each player, and keep
track of the message for that player. When the stats come back, they will be
formatted and posted to the same message as an update.