Add RTM auto restarting logic
RTM keeps closing for unknown reasons. Instead of debugging BotKit which might
take me longer than I want to spend on this, I made a start_rtm function that is
run once at the start of the main method and is retriggered whenever the
controller gets an RTM Close event.